Refactor common modules
